pyproject.toml
[project]
name = "mcpserverdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.3",
    "nest-asyncio>=1.6.0",
]
